+++ src/main/java/com/yapp/pet/web/comment/CommentController.java
@@ -1,15 +1,13 @@
 package com.yapp.pet.web.comment;
 
 import com.yapp.pet.domain.account.entity.Account;
-import com.yapp.pet.domain.comment.CommentService;
+import com.yapp.pet.domain.comment.service.CommentService;
 import com.yapp.pet.global.annotation.AuthAccount;
 import com.yapp.pet.web.comment.model.CommentRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
